cambios de ui en el perfil de la tienda
