+++ mobile/app/src/main/java/com/mylook/mylook/storeProfile/StoreContactFragment.java
@@ -16,18 +16,27 @@ import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import com.bumptech.glide.Glide;
 import com.google.common.base.Strings;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.FirebaseFirestore;
 import com.mylook.mylook.R;
+import com.mylook.mylook.entities.Subscription;
+import com.mylook.mylook.session.Session;
+
+import de.hdodenhof.circleimageview.CircleImageView;
 
 public class StoreContactFragment extends Fragment {
 
     private LinearLayout lnlface,lnlInsta,lnlTw;
     private TextView storeLocation, storePhone;
-    private TextView txtTwitter, txtInstagram;
-    private TextView txtFacebook;
-    private ImageButton btnReturnStoreInfo;
     private TextView lblMaps;
+    private CircleImageView storePhoto;
+    private Button btnSubscribe;
+    private boolean mSubscribed;
+    private String subscriptionDocument;
 
 
     public StoreContactFragment() {
@@ -44,6 +53,8 @@ public class StoreContactFragment extends Fragment {
         Bundle args = getArguments();
 
         if (args != null) {
+            String store = args.getString("name");
+
             lblMaps = rootView.findViewById(R.id.txtMaps);
             setOnClickLocation(args.getString("storeName"), args.getDouble("latitude"), args.getDouble("longitude"));
 
@@ -54,20 +65,22 @@ public class StoreContactFragment extends Fragment {
             setStorePhone(args.getString("phone"));
 
             lnlface=rootView.findViewById(R.id.lnlFace);
-            txtFacebook =rootView.findViewById(R.id.txtFacebook);
             setOnClickFacebook(args.getString("facebook"));
 
             lnlTw=rootView.findViewById(R.id.lnlTw);
-            txtTwitter =rootView.findViewById(R.id.txtTwitter);
             setOnClickTwitter(args.getString("twitter"));
 
             lnlInsta=rootView.findViewById(R.id.lnlInta);
-            txtInstagram =rootView.findViewById(R.id.txtInstagram);
             setOnClickInstagram(args.getString("instagram"));
+
+            storePhoto = rootView.findViewById(R.id.premium_profile_photo);
+            setStorePhoto(args.getString("photo"));
+
+            btnSubscribe = rootView.findViewById(R.id.btn_subscribe);
+            checkFollow(store);
+            setOnClickSubscribe(store);
         }
 
-        btnReturnStoreInfo = rootView.findViewById(R.id.btnReturnStoreInfo);
-        btnReturnStoreInfo.setOnClickListener(v -> ((StoreActivity) getActivity()).returnToStoreInfo());
     }
 
     private void setStoreLocation(String storeLocation) {
@@ -77,11 +90,14 @@ public class StoreContactFragment extends Fragment {
     private void setStorePhone(String storePhone) {
         this.storePhone.setText(storePhone);
     }
+    private void setStorePhoto(String storePhoto) {
+        Glide.with(getContext()).load(storePhoto).into(this.storePhoto);
+    }
 
     private void setOnClickFacebook(String txtFacebook) {
         if (!Strings.isNullOrEmpty(txtFacebook)) {
             final String finalTxtFacebook = "https://www.facebook.com/"+txtFacebook;
-            this.txtFacebook.setOnClickListener(v -> {
+            this.lnlface.setOnClickListener(v -> {
                 Intent intent=new Intent(Intent.ACTION_VIEW, Uri.parse(finalTxtFacebook));
                 intent.setPackage("com.facebook.android");
                 try{
@@ -97,7 +113,7 @@ public class StoreContactFragment extends Fragment {
     private void setOnClickTwitter(String txtTwitter) {
         if (!Strings.isNullOrEmpty(txtTwitter)) {
             final String finalTxtTwitter = "https://twitter.com/"+txtTwitter;
-            this.txtTwitter.setOnClickListener(v -> {
+            this.lnlTw.setOnClickListener(v -> {
                 Intent intent=new Intent(Intent.ACTION_VIEW, Uri.parse(finalTxtTwitter));
                 intent.setPackage("com.twitter.android");
                 try{
@@ -114,7 +130,7 @@ public class StoreContactFragment extends Fragment {
     private void setOnClickInstagram(String txtInstagram) {
         if (!Strings.isNullOrEmpty(txtInstagram)) {
             final String finalTxtInstagram = "https://www.instagram.com/"+txtInstagram;
-            this.txtInstagram.setOnClickListener(v -> {
+            this.lnlInsta.setOnClickListener(v -> {
                 Intent intent=new Intent(Intent.ACTION_VIEW, Uri.parse(finalTxtInstagram));
                 intent.setPackage("com.instagram.android");
                 try{
@@ -144,5 +160,72 @@ public class StoreContactFragment extends Fragment {
                 startActivity(new Intent(Intent.ACTION_VIEW, gmmIntentUri));
             }
         });
+    }
+
+    private void checkFollow(String storeName) {
+        btnSubscribe.setEnabled(false);
+        try {
+            FirebaseFirestore.getInstance().collection("subscriptions")
+                    .whereEqualTo("userId", FirebaseAuth.getInstance().getCurrentUser().getUid())
+                    .whereEqualTo("storeName", storeName)
+                    .get()
+                    .addOnCompleteListener(task -> {
+                        if (task.isSuccessful() && !task.getResult().isEmpty()) {
+                            subscriptionDocument = task.getResult().getDocuments().get(0).getId();
+                            mSubscribed = true;
+                        }
+                        setupButtonSubscribe(mSubscribed);
+                        btnSubscribe.setEnabled(true);
+                    });
+        }catch (Exception e)
+        {
+            Log.e("STORE INFO", "Error en el checkeo del follow");
+        }
+    }
+
+    private void setOnClickSubscribe(String storeName) {
+        btnSubscribe.setOnClickListener(view -> {
+            btnSubscribe.setEnabled(false);
+            if (!mSubscribed) {
+                Subscription newSubscription = new Subscription(storeName, FirebaseAuth.getInstance().getCurrentUser().getUid());
+                FirebaseFirestore.getInstance().collection("subscriptions").add(newSubscription).addOnSuccessListener(documentReference -> {
+                    Log.d("Firestore task", "DocumentSnapshot written with ID: " + documentReference.getId());
+                    subscriptionDocument = documentReference.getId();
+                    setupButtonSubscribe(true);
+                    btnSubscribe.setEnabled(true);
+                    displayMessage("Ahora estás suscripto a " + storeName);
+                    Session.getInstance().updateActivitiesStatus(Session.HOME_FRAGMENT);
+                }).addOnFailureListener(e -> {
+                    Log.w("Firestore task", "Error adding document", e);
+                    btnSubscribe.setEnabled(true);
+                });
+            } else {
+                FirebaseFirestore.getInstance().collection("subscriptions").document(subscriptionDocument).delete().addOnSuccessListener(task -> {
+                    mSubscribed = false;
+                    setupButtonSubscribe(false);
+                    btnSubscribe.setEnabled(true);
+                    subscriptionDocument = null;
+                    displayMessage("Ya no estás suscripto a " + storeName);
+                    Session.getInstance().updateActivitiesStatus(Session.HOME_FRAGMENT);
+                    btnSubscribe.setEnabled(true);
+                });
+            }
+        });
+    }
+
+    private void setupButtonSubscribe(boolean subscribed) {
+        if (subscribed) {
+            btnSubscribe.setText("Desuscribirse");
+            btnSubscribe.setBackgroundColor(getResources().getColor(R.color.primary_dark));
+            mSubscribed = true;
+        } else {
+            btnSubscribe.setText("Suscribirse");
+            btnSubscribe.setBackgroundColor(getResources().getColor(R.color.accent));
+            mSubscribed = false;
+        }
+    }
+
+    private void displayMessage(String message) {
+        Toast.makeText(getContext(), message, Toast.LENGTH_SHORT).show();
     }
 }
