+++ mobile/app/src/main/res/drawable/ic_share_new.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512.001dp"
+    android:height="512.001dp"
+    android:viewportWidth="512.001"
+    android:viewportHeight="512.001">
+  <path
+      android:fillColor="@color/white"
+      android:pathData="M411.827,300.523c-34.851,0.006 -65.478,17.833 -83.422,44.827l-131.501,-64.629c4.582,-17.947 4.582,-31.495 0,-49.441l131.502,-64.629c17.942,26.994 48.57,44.821 83.42,44.827c55.33,-0.01 100.164,-44.841 100.174,-100.172c-0.01,-55.333 -44.843,-100.164 -100.174,-100.174c-55.331,0.01 -100.164,44.841 -100.174,100.174c0.002,8.583 1.415,16.769 3.444,24.72l-131.502,64.629c-17.943,-26.994 -48.571,-44.821 -83.422,-44.827C44.843,155.838 0.01,200.669 0,256.002c0.01,55.331 44.843,100.162 100.174,100.172c34.851,-0.006 65.478,-17.833 83.422,-44.827l131.502,64.629c-2.03,7.95 -3.443,16.139 -3.444,24.722c0.01,55.331 44.843,100.162 100.174,100.172c55.33,-0.01 100.164,-44.841 100.174,-100.172C511.991,345.365 467.158,300.533 411.827,300.523z"/>
+</vector>
