+++ mobile/app/src/main/res/menu/store_menu.xml
@@ -3,9 +3,9 @@
     xmlns:android="http://schemas.android.com/apk/res/android">
     <item
         android:id="@+id/share_store"
-        android:icon="@android:drawable/ic_menu_share"
+        android:icon="@drawable/ic_share_new"
         android:title="Compartir"
-        app:showAsAction="never"
+        app:showAsAction="ifRoom"
         />
 
 </menu>
